Extract Msf::DBManager::Import::Nmap
MSP-11124
